requirements.txt
eclipse-zenoh
neurokit2>=0.2.4
matplotlib>=3.5.0
numpy>=1.20.0 